requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0
social-auth-app-django>=5.4.0

# Database
psycopg2-binary>=2.9.9

# API Documentation
drf-spectacular>=0.27.0

# Celery for async tasks
celery>=5.3.0
redis>=5.0.0

# Git integration
gitpython>=3.1.40

# Search
django-postgres-extra>=2.0.8

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
markdown>=3.5.0
bleach>=6.1.0

# Development
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.18.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
